services/extract/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic-settings==2.5.2
aiokafka==0.11.0
asyncpg==0.29.0
python-magic==0.4.27
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
prometheus-client==0.20.0
